ncu new path

+++ mlwd/torch_profiler_collect.py
@@ -49,7 +49,7 @@ def main():
         tokenizer.pad_token = tokenizer.eos_token
 
     model = AutoModelForCausalLM.from_pretrained(
-        args.model, torch_dtype=torch_dtype, device_map="cuda", trust_remote_code=True
+        args.model, dtype=torch_dtype, device_map="cuda", trust_remote_code=True
     )
     model.eval()
     print("Model loaded.\n")
